+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Support arbitrary parameters to DocusignRest::Client#get_combined_document_from_envelope (Coley Brown)
 * Support event notifications in DocusignRest::Client#get_combined_document_from_envelope (Maxime Orefice)
 * Support wet_sign option on DocusignRest::Client#create_envelope_from_document (Sergio Cambra)
+* Support signHereTabs on DocusignRest::Client#get_inline_signers (Chris Sturm)
 
 ### Misc:
 * Replace monkeypatch with argument usage (Jean-Philippe Moal)